requirements.txt
requests>=2.31.0
pandas>=2.1.0
pyarrow>=14.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.5.0
bertopic>=0.16.0
umap-learn>=0.5.5
hdbscan>=0.8.33
plotly>=5.18.0
tqdm>=4.66.0
